apply no_scroll fix to SpinButtons on thermals page
