+++ lact-gui/src/app/pages/oc_page/clocks_frame/adjustment_row.rs
@@ -1,14 +1,10 @@
 use crate::{
     APP_BROKER, I18N,
-    app::{msg::AppMsg, pages::oc_adjustment::OcAdjustment},
+    app::{ext::make_event_controller_no_scroll, msg::AppMsg, pages::oc_adjustment::OcAdjustment},
 };
 use gtk::{
-    glib::{
-        SignalHandlerId,
-        object::{Cast, ObjectExt},
-        types::StaticType,
-    },
-    prelude::{AdjustmentExt, EventControllerExt, OrientableExt, RangeExt, ScaleExt, WidgetExt},
+    glib::{SignalHandlerId, object::ObjectExt},
+    prelude::{AdjustmentExt, OrientableExt, RangeExt, ScaleExt, WidgetExt},
 };
 use i18n_embed_fl::fl;
 use lact_schema::request::ClockspeedType;
@@ -192,39 +188,4 @@ impl ClockAdjustmentRow {
     pub fn get_raw_value(&self) -> i32 {
         (self.adjustment.value() / self.value_ratio) as i32
     }
-}
-
-fn make_event_controller_no_scroll() -> gtk::EventControllerScroll {
-    let controller = gtk::EventControllerScroll::new(
-        gtk::EventControllerScrollFlags::VERTICAL | gtk::EventControllerScrollFlags::HORIZONTAL,
-    );
-    controller.connect_scroll(|controller, dx, dy| {
-        if let Some(parent) = controller
-            .widget()
-            .and_then(|widget| widget.ancestor(gtk::ScrolledWindow::static_type()))
-        {
-            let scrolled_window = parent.downcast::<gtk::ScrolledWindow>().unwrap();
-
-            if dy != 0.0 {
-                let current = scrolled_window.vadjustment().value();
-                let step = scrolled_window.vadjustment().step_increment();
-
-                // This is a bit of a hack, fractional values are generally touchpad inputs (in pixels),
-                // while whole values are scroll wheel events (which should use the `step` value)
-                // With newer GTK this should be changed to getting `unit()` from the scroll controller
-                let delta = if dy.fract() == 0.0 { dy * step } else { dy };
-                scrolled_window.vadjustment().set_value(current + delta);
-            }
-
-            if dx != 0.0 {
-                let current = scrolled_window.hadjustment().value();
-                let step = scrolled_window.hadjustment().step_increment();
-                let delta = if dx.fract() == 0.0 { dy * step } else { dy };
-                scrolled_window.hadjustment().set_value(current + delta);
-            }
-        }
-
-        gtk::glib::Propagation::Stop
-    });
-    controller
 }
